♻️ Refactor gas and fee functions
