+++ src/signingClient.ts
@@ -2,26 +2,18 @@ import BigNumber from 'bignumber.js';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { EncodeObject, OfflineSigner } from '@cosmjs/proto-signing';
 import {
-  assertIsDeliverTxSuccess,
   SigningStargateClient,
   Coin,
-  SignerData,
   StdFee,
   DeliverTxResponse,
 } from '@cosmjs/stargate';
 import { ClassConfig } from '@likecoin/iscn-message-types/dist/likenft/class_data';
 import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
-import jsonStringify from 'fast-json-stable-stringify';
 
 import {
-  ISCN_REGISTRY_NAME,
-  GAS_ESTIMATOR_INTERCEPT,
-  GAS_ESTIMATOR_SLOPE,
-  GAS_ESTIMATOR_BUFFER_RATIO,
   DEFAULT_RPC_ENDPOINT,
   DEFAULT_GAS_PRICE_NUMBER,
   COSMOS_DENOM,
-  STUB_WALLET,
   ISCN_CHANGE_OWNER_GAS,
   SEND_NFT_GAS,
   LIKENFT_MINT_NFT_GAS,
@@ -39,7 +31,6 @@ import {
   ISCNSignOptions, ISCNSignPayload, MintNFTData, NewNFTClassData,
 } from './types';
 import {
-  formatISCNPayload,
   formatMsgChangeIscnRecordOwnership,
   formatMsgCreateIscnRecord,
   formatMsgUpdateIscnRecord,
@@ -56,30 +47,10 @@ import {
   formatSendAuthorizationMsgGrant,
   formatSendAuthorizationMsgRevoke,
 } from './messages/authz';
+import signOrBroadcast from './transactions/sign';
+import { estimateISCNTxFee, estimateISCNTxGas } from './transactions/iscn';
+import formatGasFee from './transactions/gas';
 
-function formatGasFee({
-  gas = DEFAULT_MESSAGE_GAS,
-  gasPrice = DEFAULT_GAS_PRICE_NUMBER,
-  gasMultiplier = 1,
-  denom,
-}: {
-  gas?: number | string,
-  gasPrice?: number | string,
-  gasMultiplier?: number | string,
-  denom: string,
-}): StdFee {
-  const fee = {
-    amount: [{
-      amount: new BigNumber(gas)
-        .multipliedBy(gasPrice)
-        .multipliedBy(gasMultiplier)
-        .toFixed(0, 0),
-      denom,
-    }],
-    gas: new BigNumber(gas).multipliedBy(gasMultiplier).toFixed(0, 0),
-  };
-  return fee;
-}
 export class ISCNSigningClient {
   private signingClient: SigningStargateClient | null = null;
 
@@ -120,7 +91,7 @@ export class ISCNSigningClient {
     await this.setSigner(signer);
   }
 
-  async fetchISCNFeeDenom() {
+  async fetchISCNFeeDenom(): Promise<void> {
     const feePerByte = await this.queryClient.queryFeePerByte();
     if (feePerByte?.denom) this.denom = feePerByte.denom;
   }
@@ -132,7 +103,7 @@ export class ISCNSigningClient {
   async esimateISCNTxGasAndFee(
     payload: ISCNSignPayload,
     { gasPrice, memo }: { gasPrice?: number, memo?: string } = {},
-  ) {
+  ): Promise<{ gas: { fee: StdFee; }; iscnFee: Coin; }> {
     const [fee, iscnFee] = await Promise.all([
       this.estimateISCNTxGas(payload, { gasPrice, memo }),
       this.estimateISCNTxFee(payload),
@@ -146,125 +117,16 @@ export class ISCNSigningClient {
   async estimateISCNTxFee(
     payload: ISCNSignPayload,
     { version = 1 } = {},
-  ) {
-    const record = formatISCNPayload(payload);
-    const feePerByte = await this.queryClient.queryFeePerByte();
-    const feePerByteAmount = feePerByte ? parseInt(feePerByte.amount, 10) : 1;
-    const {
-      recordNotes,
-      contentFingerprints,
-      stakeholders,
-      contentMetadata,
-    } = record;
-    const now = new Date();
-    const obj = {
-      '@context': {
-        '@vocab': 'http://iscn.io/',
-        recordParentIPLD: {
-          '@container': '@index',
-        },
-        stakeholders: {
-          '@context': {
-            '@vocab': 'http://schema.org/',
-            entity: 'http://iscn.io/entity',
-            rewardProportion: 'http://iscn.io/rewardProportion',
-            contributionType: 'http://iscn.io/contributionType',
-            footprint: 'http://iscn.io/footprint',
-          },
-        },
-        contentMetadata: {
-          '@context': null,
-        },
-      },
-      '@type': 'Record',
-      '@id': `iscn://${ISCN_REGISTRY_NAME}/btC7CJvMm4WLj9Tau9LAPTfGK7sfymTJW7ORcFdruCU/1`,
-      recordTimestamp: now.toISOString(),
-      recordVersion: version,
-      recordNotes,
-      contentFingerprints,
-      recordParentIPLD: {},
-    };
-    if (version > 1) {
-      obj.recordParentIPLD = {
-        '/': 'bahuaierav3bfvm4ytx7gvn4yqeu4piiocuvtvdpyyb5f6moxniwemae4tjyq',
-      };
-    }
-    const byteSize = Buffer.from(jsonStringify(obj), 'utf-8').length
-      + Buffer.from(jsonStringify({ stakeholders: [], contentMetadata: {} }), 'utf-8').length
-      + stakeholders.reduce((acc, s) => acc + s.length, 0)
-      + stakeholders.length
-      + contentMetadata.length;
-    const feeAmount = new BigNumber(byteSize).multipliedBy(feePerByteAmount);
-    return {
-      amount: feeAmount.toFixed(0, 0),
-      denom: feePerByte?.denom || this.denom,
-    } as Coin;
+  ): Promise<Coin> {
+    return estimateISCNTxFee(this.queryClient, payload, this.denom, { version });
   }
 
-  async estimateISCNTxGas(payload: ISCNSignPayload, {
-    denom = this.denom,
-    gasPrice = DEFAULT_GAS_PRICE_NUMBER,
-    memo,
-  }: {
-    denom?: string,
-    gasPrice?: number,
-    memo?: string,
-  } = {}) {
-    const msg = formatMsgCreateIscnRecord(STUB_WALLET, payload);
-    const value = {
-      msg: [msg],
-      // temp number here for estimation
-      fee: formatGasFee({ gas: '200000', gasPrice: '1', denom }),
-    };
-    const obj = {
-      type: 'cosmos-sdk/StdTx',
-      value,
-      memo, // directly append memo to object if exists, since we only need its length
-    };
-    const txBytes = Buffer.from(jsonStringify(obj), 'utf-8');
-    const byteSize = new BigNumber(txBytes.length);
-    const gasUsedEstimationBeforeBuffer = byteSize
-      .multipliedBy(GAS_ESTIMATOR_SLOPE)
-      .plus(GAS_ESTIMATOR_INTERCEPT);
-    const buffer = gasUsedEstimationBeforeBuffer.multipliedBy(GAS_ESTIMATOR_BUFFER_RATIO);
-    const gasUsedEstimation = gasUsedEstimationBeforeBuffer.plus(buffer);
-    const gas = gasUsedEstimation.toFixed(0, 0);
-    return formatGasFee({ gas, gasPrice, denom });
-  }
-
-  private async signOrBroadcast(
-    senderAddress: string,
-    messages: EncodeObject[],
-    fee: StdFee,
-    {
-      memo = '',
-      broadcast = true,
-      accountNumber,
-      sequence,
-      chainId,
-    }: ISCNSignOptions = {},
-  ) {
-    const client = this.signingClient;
-    if (!client) throw new Error('SIGNING_CLIENT_NOT_CONNECTED');
-    let signData: SignerData | undefined;
-    if ((accountNumber !== undefined || sequence !== undefined || chainId !== undefined)) {
-      if (!(accountNumber !== undefined && sequence !== undefined && chainId !== undefined)) {
-        throw new Error('MUST_DEFINE_ALL_SIGNING_PARAM');
-      }
-      signData = {
-        accountNumber,
-        sequence,
-        chainId,
-      };
-    }
-    const txRaw = await client.sign(senderAddress, messages, fee, memo, signData);
-    if (!broadcast) {
-      return txRaw;
-    }
-    const txBytes = TxRaw.encode(txRaw).finish();
-    const response = await client.broadcastTx(txBytes);
-    assertIsDeliverTxSuccess(response);
-    return response;
+  estimateISCNTxGas(
+    payload: ISCNSignPayload,
+    { denom = this.denom, gasPrice = DEFAULT_GAS_PRICE_NUMBER, memo }
+    : { denom?: string, gasPrice?: number, memo?: string } = {},
+  ): StdFee {
+    return estimateISCNTxGas(payload, { denom, gasPrice, memo });
   }
 
   async sendMessages(
@@ -281,7 +143,7 @@ export class ISCNSigningClient {
       gasMultiplier: messages.length,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -292,14 +154,14 @@ export class ISCNSigningClient {
   ): Promise<TxRaw | DeliverTxResponse> {
     const client = this.signingClient;
     if (!client) throw new Error('SIGNING_CLIENT_NOT_CONNECTED');
-    const message = formatMsgCreateIscnRecord(senderAddress, payload);
+    const messages = [formatMsgCreateIscnRecord(senderAddress, payload)];
     let fee = inputFee;
     if (fee && gasPrice) throw new Error('CANNOT_SET_BOTH_FEE_AND_GASPRICE');
     if (!fee) {
       const { memo } = signOptions;
-      fee = await this.estimateISCNTxGas(payload, { gasPrice, memo });
+      fee = this.estimateISCNTxGas(payload, { gasPrice, memo });
     }
-    const response = await this.signOrBroadcast(senderAddress, [message], fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -311,14 +173,14 @@ export class ISCNSigningClient {
   ): Promise<TxRaw | DeliverTxResponse> {
     const client = this.signingClient;
     if (!client) throw new Error('SIGNING_CLIENT_NOT_CONNECTED');
-    const message = formatMsgUpdateIscnRecord(senderAddress, iscnId, payload);
+    const messages = [formatMsgUpdateIscnRecord(senderAddress, iscnId, payload)];
     let fee = inputFee;
     if (fee && gasPrice) throw new Error('CANNOT_SET_BOTH_FEE_AND_GASPRICE');
     if (!fee) {
       const { memo } = signOptions;
       fee = await this.estimateISCNTxGas(payload, { gasPrice, memo });
     }
-    const response = await this.signOrBroadcast(senderAddress, [message], fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -330,14 +192,14 @@ export class ISCNSigningClient {
   ): Promise<TxRaw | DeliverTxResponse> {
     const client = this.signingClient;
     if (!client) throw new Error('SIGNING_CLIENT_NOT_CONNECTED');
-    const message = formatMsgChangeIscnRecordOwnership(senderAddress, iscnId, newOwnerAddress);
+    const messages = [formatMsgChangeIscnRecordOwnership(senderAddress, iscnId, newOwnerAddress)];
     if (inputFee && gasPrice) throw new Error('CANNOT_SET_BOTH_FEE_AND_GASPRICE');
     const fee = inputFee || formatGasFee({
       gas: ISCN_CHANGE_OWNER_GAS,
       gasPrice,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, [message], fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -365,19 +227,19 @@ export class ISCNSigningClient {
         ...(nftClassData.metadata || {}),
       },
     };
-    const message = formatMsgNewClass(
+    const messages = [formatMsgNewClass(
       senderAddress,
       iscnIdPrefix,
       combinedClassData,
       classConfig,
-    );
+    )];
     if (inputFee && gasPrice) throw new Error('CANNOT_SET_BOTH_FEE_AND_GASPRICE');
     const fee = inputFee || formatGasFee({
       gas: LIKENFT_CREATE_CLASS_GAS,
       gasPrice,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, [message], fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -390,19 +252,19 @@ export class ISCNSigningClient {
   ): Promise<TxRaw | DeliverTxResponse> {
     const client = this.signingClient;
     if (!client) throw new Error('SIGNING_CLIENT_NOT_CONNECTED');
-    const message = formatMsgUpdateClass(
+    const messages = [formatMsgUpdateClass(
       senderAddress,
       classId,
       nftClassData,
       classConfig,
-    );
+    )];
     if (inputFee && gasPrice) throw new Error('CANNOT_SET_BOTH_FEE_AND_GASPRICE');
     const fee = inputFee || formatGasFee({
       gas: LIKENFT_UPDATE_CLASS_GAS,
       gasPrice,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, [message], fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -425,7 +287,7 @@ export class ISCNSigningClient {
       gasMultiplier: nftDatas.length,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -446,7 +308,7 @@ export class ISCNSigningClient {
       gasMultiplier: nftIds.length,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -466,7 +328,7 @@ export class ISCNSigningClient {
       gasMultiplier: nftIds.length,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -498,7 +360,7 @@ export class ISCNSigningClient {
     } else if (gasPrice) {
       throw new Error('CANNOT_SET_BOTH_FEE_AND_GASPRICE');
     }
-    const response = await this.signOrBroadcast(senderAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -523,7 +385,7 @@ export class ISCNSigningClient {
       gasPrice,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(execAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(execAddress, messages, fee, client, signOptions);
     return response;
   }
 
@@ -541,7 +403,7 @@ export class ISCNSigningClient {
       gasPrice,
       denom: this.denom,
     });
-    const response = await this.signOrBroadcast(senderAddress, messages, fee, signOptions);
+    const response = await signOrBroadcast(senderAddress, messages, fee, client, signOptions);
     return response;
   }
 }
